chore(test): separate test package structure
