+++ apps/backend/src/test/java/scrumpledpaper/agiler/AgilerApplicationTests.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 
+import scrumpledpaper.agiler.config.TestcontainersConfiguration;
+
 @Import(TestcontainersConfiguration.class)
 @SpringBootTest
 class AgilerApplicationTests {
