+++ apps/backend/src/test/java/scrumpledpaper/agiler/TestAgilerApplication.java
@@ -2,6 +2,8 @@ package scrumpledpaper.agiler;
 
 import org.springframework.boot.SpringApplication;
 
+import scrumpledpaper.agiler.config.TestcontainersConfiguration;
+
 public class TestAgilerApplication {
 
 	public static void main(String[] args) {
