+++ apps/backend/src/test/java/scrumpledpaper/agiler/config/TestcontainersConfiguration.java
@@ -1,4 +1,4 @@
-package scrumpledpaper.agiler;
+package scrumpledpaper.agiler.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
